Replaced Parquet FileStream with FileSink.
