+++ src/io/parquet/write/mod.rs
@@ -9,7 +9,6 @@ mod primitive;
 mod row_group;
 mod schema;
 mod sink;
-mod stream;
 mod utf8;
 mod utils;
 
@@ -41,7 +40,6 @@ pub use file::FileWriter;
 pub use row_group::{row_group_iter, RowGroupIterator};
 pub use schema::to_parquet_type;
 pub use sink::FileSink;
-pub use stream::FileStreamer;
 
 pub(self) fn decimal_length_from_precision(precision: usize) -> usize {
     // digits = floor(log_10(2^(8*n - 1) - 1))
